JSDK-3018 Moving twilio-video.d.ts contents into index.d.ts
