+++ tsdef/index.d.ts
@@ -1,3 +1,15 @@
+import { ConnectOptions, CreateLocalTrackOptions, CreateLocalTracksOptions, LocalTrack } from './types';
+import { LocalAudioTrack } from './LocalAudioTrack';
+import { LocalVideoTrack } from './LocalVideoTrack';
+import { Room } from './Room';
+
+export const isSupported: boolean;
+export const version:string;
+export function connect(token: string, options?: ConnectOptions): Promise<Room>;
+export function createLocalAudioTrack(options?: CreateLocalTrackOptions): Promise<LocalAudioTrack>;
+export function createLocalTracks(options?: CreateLocalTracksOptions): Promise<LocalTrack[]>;
+export function createLocalVideoTrack(options?: CreateLocalTrackOptions): Promise<LocalVideoTrack>;
+
 export { AudioTrack } from './AudioTrack';
 export { LocalTrack, RemoteTrack } from './types';
 export { LocalAudioTrack } from './LocalAudioTrack';
@@ -22,4 +34,3 @@ export { Track } from './Track';
 export { TrackPublication } from './TrackPublication';
 export { TwilioError } from './TwilioError';
 export { VideoTrack } from './VideoTrack';
-export * from './twilio-video';
